last step

+++ src/main/java/com/mycompany/rest/webservices/restfulwebservices/todo/TodoHardcodedService.java
@@ -1,16 +1,16 @@
 package com.mycompany.rest.webservices.restfulwebservices.todo;
 
-import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+
+import org.springframework.stereotype.Service;
 
 
 @Service
 public class TodoHardcodedService {
     private static List<Todo> todos = new ArrayList();
-    private static int idCounter = -1;
+    private static long idCounter = -1;
 
 
     static {
